update `core_function_count` in more cases; update wasm-tools (#18)
This is related to porting `componentize-py` to use `wit-dylib`.

Note that `component-init` is mostly obsolete at this point now that Wizer has
component support.  For now, though, `componentize-py` has to keep using it due
to how `--stub-wasi` is implemented so as to work around `wasmtime-py`
limitations.  Alex is working to remove those limitations, at which point we'll
be able to remove `--stub-wasi` and lean on e.g. `wasi-virt` instead.  In the
meantime, `component-init` exists only for `componentize-py`, so I'll treat it
that way :)

Signed-off-by: Joel Dice <joel.dice@fermyon.com>